refactor: move admin stats adapter to infra
