+++ crates/dcc-mcp-gateway/src/gateway/admin/mod.rs
@@ -112,7 +112,6 @@ mod skill_paths;
 pub mod skill_reload;
 pub mod sqlite_lane;
 pub mod state;
-pub mod stats;
 pub mod trace;
 #[cfg(feature = "admin")]
 mod traffic;
@@ -157,13 +156,13 @@ pub use dcc_mcp_db::{
     default_gateway_admin_sqlite_path as default_admin_db_path,
     resolve_gateway_admin_sqlite_path as resolve_admin_db_path,
 };
-pub use infra::audit_sink::AdminAuditSink;
+pub use infra::stats;
+pub use infra::{
+    AdminAuditSink, GatewayStats, LatencyStats, StatsAggregator, StatsFilter, StatsRange,
+    StatsStatus, TopEntry,
+};
 pub use sqlite_lane::{AdminSqliteLane, AdminSqliteReader, read_custom_skill_paths_for_startup};
 pub use state::{AdminAuditRecord, AdminState, AuditLog, DurableAuditStore};
-pub use stats::{
-    GatewayStats, LatencyStats, StatsAggregator, StatsFilter, StatsRange, StatsStatus, TopEntry,
-};
-
 pub use trace::{DispatchTrace, TraceContext, TraceLog, TracePayload, TraceSpan};
 
 #[cfg(feature = "admin")]
